+++ lib/solar_test/src/benchmarks/regression_detector.cpp
@@ -366,7 +366,7 @@ void RegressionDetector::add_trend_data_point(const std::string& benchmark_name,
   // Keep only recent data points (e.g., last 100)
   auto& data = trend_data_[benchmark_name];
   if (data.size() > 100) {
-    data.erase(data.begin(), data.begin() + (data.size() - 100));
+    data.erase(data.begin(), data.begin() + static_cast<std::ptrdiff_t>(data.size() - 100));
   }
 
   save_trend_data();
@@ -426,8 +426,8 @@ std::string RegressionDetector::generate_regression_report(
       << "\n\n";
 
   size_t total_benchmarks = analyses.size();
-  size_t regressions = std::count_if(analyses.begin(), analyses.end(),
-                                     [](const auto& a) { return a.has_regression; });
+  size_t regressions = static_cast<size_t>(std::count_if(analyses.begin(), analyses.end(),
+                                     [](const auto& a) { return a.has_regression; }));
 
   oss << "Summary:\n";
   oss << "  Total Benchmarks: " << total_benchmarks << "\n";
@@ -468,8 +468,8 @@ void RegressionDetector::generate_ci_output(const std::vector<RegressionAnalysis
   file << "  \"regression_summary\": {\n";
   file << "    \"total_benchmarks\": " << analyses.size() << ",\n";
 
-  size_t regressions = std::count_if(analyses.begin(), analyses.end(),
-                                     [](const auto& a) { return a.has_regression; });
+  size_t regressions = static_cast<size_t>(std::count_if(analyses.begin(), analyses.end(),
+                                     [](const auto& a) { return a.has_regression; }));
   file << "    \"regressions_found\": " << regressions << ",\n";
   file << "    \"has_critical_regressions\": "
        << (has_critical_regressions(analyses) ? "true" : "false") << "\n";
